requirements.txt
asgiref==3.8.1
Django==4.2.11
et-xmlfile==1.1.0
kaleido==0.2.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
simpleui==4.0.2
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.11.0
tzdata==2024.1
xlrd==2.0.1
